requirements.txt
streamlit
google-generativeai
python-dotenv
elevenlabs
moviepy<2.0.0
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
Pillow
